Bugfixes on backend, city component on frontend, some styling

+++ MediApp/Application/CommandsAndQueries/Cities/Queries/GetCityDetails/CityDetailsVm.cs
@@ -9,7 +9,7 @@ namespace Application.CommandsAndQueries
     public class CityDetailsVm : IMapFrom<City>
     {
         public string Name { get; set; }
-        public short CountyId { get; set; }
+        public int CountyId { get; set; }
         public bool? Deleted { get; set; }
     }
 }
